+++ src/model/mod.rs
@@ -7,6 +7,7 @@ mod items;
 mod mapping;
 mod row_col;
 mod transaction;
+mod transaction_ids;
 
 pub use amount::{Amount, AmountFormat};
 pub use auto_cat::{AutoCat, AutoCatUpdates, AutoCats};
@@ -18,6 +19,7 @@ pub(crate) use mapping::Mapping;
 pub(crate) use row_col::RowCol;
 use serde::{Deserialize, Serialize};
 pub use transaction::{Transaction, TransactionColumn, TransactionUpdates, Transactions};
+pub(crate) use transaction_ids::resolve_transaction_ids;
 
 /// Represents all the sheets of interest from a tiller Google sheet.
 #[derive(Default, Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]